Started writing ARMCount pass.
Started writing the ARMCount pass for the PicoXOM project and
added it to the default pipeline for ARM code generation.
This pass calculates statistics about the code size which we can
use to determine the amount of entropy of various code layout
randomization strategies.

The current commit calcultes the number of locations in which a
function can be placed within the code segment space of an ARM
embedded systems board.